Merge NDVs when computing statistics for ORs
Consider the OR predicate (a = 10) OR (a = 15) and the input histogram has
only NDVs. The statistics function computes result histogram for (a = 10)
as well as (a = 15) independently and then merges the two resultant
histogram in one.

The merge operation was too strict and took the maximum of the number of
distinct values (NDV)  of the two histograms to be the value for the
merged histogram. The underlying reason is correct for predicates such
as (a = 10) OR (a < 15) since this logic avoids double counting.

However, for equality operation adding up the two NDVs is fine.

It is important to note that duplicate predicates are removed during the
preprocessing step. The test added is an array comparison which is
converted into an OR before cardinality estimation